CASS-56 | Create connection buillder to make code cleaner
