+++ src/main/java/org/mule/modules/cassandradb/internal/connection/CassandraConnection.java
@@ -25,6 +25,11 @@ public final class CassandraConnection implements ConnectorConnection {
 
     private static final Logger logger = LoggerFactory.getLogger(CassandraConnection.class);
 
+    public CassandraConnection(Cluster cluster, Session session) {
+        this.cluster = cluster;
+        this.cassandraSession = session;
+    }
+
     /**
      * Connect to Cassandra Cluster specified by provided host IP
      * address and port number.
@@ -33,64 +38,8 @@ public final class CassandraConnection implements ConnectorConnection {
      * @return CassandraConnection created
      * @throws CassandraException if any error occurs when trying to connect
      */
-    public static CassandraConnection buildCassandraClient(ConnectionParameters connectionParameters) throws CassandraException {
-        validateBasicParams(connectionParameters);
-        Cluster.Builder clusterBuilder;
-        try {
-            clusterBuilder = Cluster.builder().addContactPoint(connectionParameters.getHost()).withPort(Integer.parseInt(connectionParameters.getPort()));
-        } catch (IllegalArgumentException connEx) {
-            logger.error("Error while connecting to Cassandra database!", connEx);
-            throw new CassandraException(connEx.getMessage());
-        }
-
-        if (StringUtils.isNotEmpty(connectionParameters.getUsername()) && StringUtils.isNotEmpty(connectionParameters.getPassword())) {
-            clusterBuilder.withCredentials(connectionParameters.getUsername(), connectionParameters.getPassword());
-        }
-
-        if (connectionParameters.getAdvancedConnectionParameters() != null) {
-            addAdvancedConnectionParameters(clusterBuilder, connectionParameters.getAdvancedConnectionParameters());
-        }
-
-        CassandraConnection client = new CassandraConnection();
-        client.cluster = clusterBuilder.build();
-
-        try {
-            logger.info("Connecting to Cassandra Database: {} , port: {} with clusterName: {} , protocol version {} and compression type {} ",
-                    connectionParameters.getHost(),
-                    connectionParameters.getPort(),
-                    connectionParameters.getAdvancedConnectionParameters() != null ? connectionParameters.getAdvancedConnectionParameters().getClusterName() : null,
-                    connectionParameters.getAdvancedConnectionParameters() != null ? connectionParameters.getAdvancedConnectionParameters().getProtocolVersion() : null,
-                    connectionParameters.getAdvancedConnectionParameters() != null ? connectionParameters.getAdvancedConnectionParameters().getCompression() : null);
-            client.cassandraSession = StringUtils.isNotEmpty(connectionParameters.getKeyspace()) ? client.cluster.connect(connectionParameters.getKeyspace())
-                    : client.cluster.connect();
-            logger.info("Connected to Cassandra Cluster: {} !", client.cassandraSession.getCluster().getClusterName());
-        } catch (Exception cassandraException) {
-            logger.error("Error while connecting to Cassandra database!", cassandraException);
-            throw new CassandraException(cassandraException.getMessage());
-        }
-        return client;
-    }
-
-    private static void addAdvancedConnectionParameters(Cluster.Builder clusterBuilder, AdvancedConnectionParameters advancedConnectionParameters) {
-        if (StringUtils.isNotEmpty(advancedConnectionParameters.getClusterName())) {
-            clusterBuilder.withClusterName(advancedConnectionParameters.getClusterName());
-        }
-
-        if (advancedConnectionParameters.getMaxSchemaAgreementWaitSeconds() > 0) {
-            clusterBuilder.withMaxSchemaAgreementWaitSeconds(advancedConnectionParameters.getMaxSchemaAgreementWaitSeconds());
-        }
-
-        if (advancedConnectionParameters.getProtocolVersion() != null) {
-            clusterBuilder.withProtocolVersion(advancedConnectionParameters.getProtocolVersion());
-        }
-
-        if (advancedConnectionParameters.getCompression() != null) {
-            clusterBuilder.withCompression(advancedConnectionParameters.getCompression());
-        }
-
-        if (advancedConnectionParameters.isSsl()) {
-            clusterBuilder.withSSL();
-        }
+    public static CassandraConnection build(ConnectionParameters connectionParameters) throws CassandraException {
+        return CassandraConnectionBuilder.build(connectionParameters);
     }
 
     private void closeCluster() {
@@ -102,12 +51,6 @@ public final class CassandraConnection implements ConnectorConnection {
     private void closeSession() {
         if (cassandraSession != null) {
             cassandraSession.close();
-        }
-    }
-
-    private static void validateBasicParams(ConnectionParameters parameters) {
-        if (StringUtils.isBlank(parameters.getHost()) || StringUtils.isBlank(parameters.getPort())) {
-            throw new IllegalArgumentException("Unable to connect! Missing HOST or PORT parameter!");
         }
     }
